nixos/binfmt: Preserve argv[0] when using QEMU
Traditionally, the emulator invoked by `binfmt-misc` doesn't know what the
original argv[0] is when running an executable, because only the absolute
path to the foreign executable is passed.

`binfmt-misc` now has the "P" flag that, when enabled, will add an extra
argument after the absolute path to the foreign binary containing the
original argv[0]. In our wrapper script, we rewrite this argument to
into what QEMU accepts.