pythonPackages.hypothesis: Disable tests on RISC-V
This is for no reason other than it is simply *too slow*.
So slow, in fact, that it fails its own healthcheck in some
tests:

> hypothesis.errors.FailedHealthCheck: Data generation is
> extremely slow: Only produced 5 valid examples in 1.01
> seconds (0 invalid ones and 1 exceeded maximum size). Try
> decreasing size of the data you're generating (with
> e.g.max_size or max_leaves parameters).

Instead of disabling individual tests, let's disable the
test suite for the time being since it's really not fun
seeing things randomly fail after 10+ hours of build.

I'm using the fastest hardware that is publicly available
at the moment.